fix typing error in commitCache.ts
I wrote this commit a while back before the error was disabled in commit
684939f4642c. I don't think it makes any difference in practice but I
guess might as well clean it up?

My understanding is that this is intended to be like a dictionary
declaration, but written as it currently is, the type checker expects
_every_ key to map to the value type.

Giving it a undefined option makes the issue go away.

Note: I'm a total noob in typescript, not sure if there is a prettier
way to write this than the`!== undefined` check?