avoid N getCommit calls when N commits are out of sync with upstream
Each getCommit call ends up as a `git show` subprocess:

https://github.com/microsoft/vscode/blob/65a3d201e6a97b56ad0980bf2078cf6381ce3e52/extensions/git/src/git.ts#L2468

In my Linux kernel repo my `master` branch goes unused for months at a
time so when I check it out it's normally tens of thousands of commits
behind its upstream. Spinning up that many `git show` commands kills the
extension! Not sure of the exact mechanics of why it gets killed, but
even if it didn't get killed, it would be very slow and power-hungry.

I initially tried to fix this by avoiding the `getCommit` calls, instead
fetching partial commit data directly from the log. This turned out to
be a bit of a pain, it required introducing an extra type to represent
that incomplete commit data. Further testing then revealed this wasn't
even very fast.

So, the new solution is just to have the command code limit the number
of commits that we fetch to 50 for each of the ahead/behind gap.